+++ util/unit_transfer.go
@@ -3,6 +3,6 @@ package util
 import "math"
 
 // ToCents 将元转为分
-func ToCents(amount float64) int32 {
-	return int32(math.Round(amount * 100))
+func ToCents(amount float64) int64 {
+	return int64(math.Round(amount * 100))
 }
